requirements.txt
asgiref==3.8.1
boto3==1.26.137
botocore==1.29.137
certifi==2023.5.7
cffi==1.17.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
dj-database-url==2.0.0
Django==5.1.4
django-cors-headers==4.6.0
django-ses==3.5.0
django-storages==1.13.2
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
oauthlib==3.2.2
psycopg==3.2.4
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2024.2
requests==2.30.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
setuptools==75.8.0
six==1.17.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.5.3

typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.15

################################################

boto3==1.26.137
botocore==1.29.137
certifi==2023.5.7
cffi==1.17.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
dj-database-url==2.0.0
dj-rest-auth==7.0.1

django-allauth==65.3.1

django-filter==24.3
django-ses==3.5.0
django-storages==1.13.2
django-templated-mail==1.1.1

djangorestframework-simplejwt==5.2.2
djoser==2.2.0
et_xmlfile==2.0.0
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
Markdown==3.7
numpy==2.2.0
oauthlib==3.2.2
openpyxl==3.1.5
pandas==2.2.3
psycopg==3.2.4
pycparser==2.21
PyJWT==2.7.0

python-dotenv==1.0.0
python3-openid==3.2.0

requests==2.30.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
setuptools==75.8.0

social-auth-app-django==5.2.0
social-auth-core==4.4.2


urllib3==1.26.15
